backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
build==1.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.2
numpy-stl==3.2.0
orjson==3.11.1
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.0
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.23.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.12
python-utils==3.9.1
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.3
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.34.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.30.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
